requirements.txt
django
tzdata
celery
requests
python-telegram-bot
asgiref
pillow==11.1.0
qrcode==8.0
